requirements.txt
pydantic
sqlalchemy
protobuf
Flask==2.3.3
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0
googleapis-common-protos==1.62.0
google-cloud-aiplatform==1.43.0   # opcional, só se usar cliente direto
requests==2.31.0
python-dotenv==1.0.0
google-genai
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.1
fastapi==0.111.0
uvicorn==0.30.1
aiosqlite
pydub

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
httpx==0.25.2